backend/auth-server/requirements.txt
fastapi
uvicorn
python-dotenv
better-auth
psycopg2-binary
